Part8 complete

+++ Infraestructure/Services/AccountService.cs
@@ -239,7 +239,7 @@ public class AccountService : IAccountService
                 Directory.CreateDirectory(uploadPath);
             }
 
-            if (currentUser.Value.Avatar != Constants.DEFAULT_AVATAR)
+            if (currentUser.Value.Avatar != null && currentUser.Value.Avatar != Constants.DEFAULT_AVATAR)
             {
                 previousAvatar = currentUser.Value.Avatar;
                 previousAvatar = Path.Combine(uploadPath, previousAvatar);
@@ -296,10 +296,14 @@ public class AccountService : IAccountService
         if (currentUser.Value.Avatar != Constants.DEFAULT_AVATAR)
         {
             previousAvatar = currentUser.Value.Avatar;
-            previousAvatar = Path.Combine(uploadPath, previousAvatar);
-            if (File.Exists(previousAvatar))
+
+            if (!string.IsNullOrEmpty(previousAvatar))
             {
-                File.Delete(previousAvatar);
+                previousAvatar = Path.Combine(uploadPath, previousAvatar);
+                if (File.Exists(previousAvatar))
+                {
+                    File.Delete(previousAvatar);
+                }
             }
 
             currentUser.Value.Avatar = Constants.DEFAULT_AVATAR;
